Removing i hyperdrive policy and moving tests
