+++ src/agent0/core/hyperdrive/interactive/i_local_hyperdrive.py
@@ -68,7 +68,6 @@ from .event_types import (
     RemoveLiquidity,
 )
 from .i_hyperdrive import IHyperdrive
-from .i_hyperdrive_policy import IHyperdrivePolicy
 from .i_local_chain import ILocalChain
 from .i_local_hyperdrive_agent import ILocalHyperdriveAgent
